fix: add missing var declarations
